REFACTOR: Move specific analyses to `stdlib` folder.
This mirrors the code organization in `catlog`.
