+++ packages/frontend/src/analysis/index.ts
@@ -1,8 +1,2 @@
 export * from "./types";
 export * from "./document";
-export * from "./analysis_editor";
-
-export * from "./lotka_volterra";
-export * from "./model_graph";
-export * from "./stock_flow_diagram";
-export * from "./submodel_graphs";
