+++ packages/frontend/src/model/index.ts
@@ -1,7 +1,3 @@
 export * from "./types";
 export * from "./context";
 export * from "./document";
-
-export * from "./object_cell_editor";
-export * from "./morphism_cell_editor";
-export * from "./model_editor";
